# openapi.yaml
openapi: 3.1.0
info:
  title: Documan API
  version: 1.0.0
  description: |
    REST API for organising academic subjects and files, and for the surrounding community
    features (posts, comments, votes, favourites, users, roles).

    This document is maintained by hand rather than generated at runtime. When an endpoint,
    payload or status code changes, update it in the same commit.

    Notes on conventions:
      * Collection endpoints are paginated and return a `PageResponse` envelope.
      * Errors use RFC 9457 `application/problem+json`.
      * Identifiers are passed as query parameters.
      * There is currently no authentication; every operation is reachable anonymously.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
  /api/v1/user:
    get:
      tags: [User]
      summary: Fetch a user by id
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      responses:
        "200": { $ref: "#/components/responses/User" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [User]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201": { $ref: "#/components/responses/User" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
    put:
      tags: [User]
      summary: Update a user
      description: A null `password` leaves the stored credential unchanged.
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserRequest" }
      responses:
        "200": { $ref: "#/components/responses/User" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [User]
      summary: Delete a user
      description: Fails with 409 if the user still owns posts or comments.
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/v1/user/all:
    get:
      tags: [User]
      summary: List users
      parameters: [{ $ref: "#/components/parameters/Pageable" }]
      responses:
        "200": { $ref: "#/components/responses/UserPage" }
...
  /api/v1/post:
    get:
      tags: [Post]
      summary: Fetch a post by id
      parameters: [{ $ref: "#/components/parameters/PostId" }]
      responses:
        "200": { $ref: "#/components/responses/Post" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Post]
      summary: Create a post
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePostRequest" }
      responses:
        "201": { $ref: "#/components/responses/Post" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Post]
      summary: Update a post
      parameters: [{ $ref: "#/components/parameters/PostId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePostRequest" }
      responses:
        "200": { $ref: "#/components/responses/Post" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [Post]
      summary: Delete a post and its comments
      parameters: [{ $ref: "#/components/parameters/PostId" }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/post/all:
    get:
      tags: [Post]
      summary: List posts, newest first
      parameters: [{ $ref: "#/components/parameters/Pageable" }]
      responses:
        "200": { $ref: "#/components/responses/PostPage" }
  /api/v1/post/user:
    get:
      tags: [Post]
      summary: List posts by author
      parameters:
        - { $ref: "#/components/parameters/UserId" }
        - { $ref: "#/components/parameters/Pageable" }
      responses:
        "200": { $ref: "#/components/responses/PostPage" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/post/vote:
    post:
      tags: [Post]
      summary: Cast or switch a vote on a post
      description: |
        Idempotent. Casting the same direction twice leaves the tallies unchanged; casting the
        opposite direction moves the vote rather than adding a second one.
      parameters:
        - { $ref: "#/components/parameters/VoteType" }
        - { $ref: "#/components/parameters/PostId" }
        - { $ref: "#/components/parameters/UserId" }
      responses:
        "200": { $ref: "#/components/responses/Post" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/post/vote/remove:
    post:
      tags: [Post]
      summary: Withdraw a vote from a post
      parameters:
        - { $ref: "#/components/parameters/VoteType" }
        - { $ref: "#/components/parameters/PostId" }
        - { $ref: "#/components/parameters/UserId" }
      responses:
        "200": { $ref: "#/components/responses/Post" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/post/favourite:
    post:
      tags: [Post]
      summary: Favourite a post
      parameters:
        - { $ref: "#/components/parameters/PostId" }
        - { $ref: "#/components/parameters/UserId" }
      responses:
        "200": { $ref: "#/components/responses/Post" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/post/favourite/remove:
    post:
      tags: [Post]
      summary: Remove a post favourite
      parameters:
        - { $ref: "#/components/parameters/PostId" }
        - { $ref: "#/components/parameters/UserId" }
      responses:
        "200": { $ref: "#/components/responses/Post" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/comment:
    get:
      tags: [Comment]
      summary: Fetch a comment by id
      parameters: [{ $ref: "#/components/parameters/CommentId" }]
      responses:
        "200": { $ref: "#/components/responses/Comment" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Comment]
      summary: Add a comment to a post
      parameters:
        - { $ref: "#/components/parameters/UserId" }
        - { $ref: "#/components/parameters/PostId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommentBody" }
      responses:
        "201": { $ref: "#/components/responses/Comment" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Comment]
      summary: Update a comment
      parameters: [{ $ref: "#/components/parameters/CommentId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommentBody" }
      responses:
        "200": { $ref: "#/components/responses/Comment" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Comment]
      summary: Delete a comment
      parameters: [{ $ref: "#/components/parameters/CommentId" }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/comment/all:
    get:
      tags: [Comment]
      summary: List comments, newest first
      parameters: [{ $ref: "#/components/parameters/Pageable" }]
      responses:
        "200": { $ref: "#/components/responses/CommentPage" }
  /api/v1/comment/user:
    get:
      tags: [Comment]
      summary: List comments by author
      parameters:
        - { $ref: "#/components/parameters/UserId" }
        - { $ref: "#/components/parameters/Pageable" }
      responses:
        "200": { $ref: "#/components/responses/CommentPage" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/comment/post:
    get:
      tags: [Comment]
      summary: List comments on a post
      parameters:
        - { $ref: "#/components/parameters/PostId" }
        - { $ref: "#/components/parameters/Pageable" }
      responses:
        "200": { $ref: "#/components/responses/CommentPage" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/comment/vote:
    post:
      tags: [Comment]
      summary: Cast or switch a vote on a comment
      parameters:
        - { $ref: "#/components/parameters/VoteType" }
        - { $ref: "#/components/parameters/CommentId" }
        - { $ref: "#/components/parameters/UserId" }
      responses:
        "200": { $ref: "#/components/responses/Comment" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/comment/vote/remove:
    post:
      tags: [Comment]
      summary: Withdraw a vote from a comment
      parameters:
        - { $ref: "#/components/parameters/VoteType" }
        - { $ref: "#/components/parameters/CommentId" }
        - { $ref: "#/components/parameters/UserId" }
      responses:
        "200": { $ref: "#/components/responses/Comment" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/subject:
    get:
      tags: [Subject]
      summary: Fetch a subject by id
      parameters: [{ $ref: "#/components/parameters/SubjectId" }]
      responses:
        "200": { $ref: "#/components/responses/Subject" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Subject]
      summary: Create a subject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubjectBody" }
      responses:
        "201": { $ref: "#/components/responses/Subject" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Subject]
      summary: Update a subject
      parameters: [{ $ref: "#/components/parameters/SubjectId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubjectBody" }
      responses:
        "200": { $ref: "#/components/responses/Subject" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Subject]
      summary: Delete a subject
      parameters: [{ $ref: "#/components/parameters/SubjectId" }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/v1/file/subject:
    get:
      tags: [File]
      summary: List files attached to a subject
      parameters:
        - { $ref: "#/components/parameters/SubjectId" }
        - { $ref: "#/components/parameters/Pageable" }
      responses:
        "200": { $ref: "#/components/responses/FilePage" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/file:
    post:
      tags: [File]
      summary: Upload a file against a subject
      description: |
        The object key is a random prefix plus a sanitised form of the original filename. If the
        metadata write fails the uploaded object is removed, so a partial failure does not leave
        an unreferenced object in the bucket.
      parameters: [{ $ref: "#/components/parameters/SubjectId" }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201": { $ref: "#/components/responses/File" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/StorageFailure" }
    delete:
      tags: [File]
      summary: Delete a file
      description: Removes both the stored object and its metadata row.
      parameters:
        - name: objectUID
          in: query
          required: true
          description: The generated object key, as returned in `objectName`.
          schema: { type: string }
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/StorageFailure" }
...
components:
  parameters:
...
    UserId:
      name: userId
      in: query
      required: true
      schema: { type: integer, format: int32 }
    PostId:
      name: postId
      in: query
      required: true
      schema: { type: integer, format: int32 }
    CommentId:
      name: commentId
      in: query
      required: true
      schema: { type: integer, format: int32 }
    SubjectId:
      name: subjectId
      in: query
      required: true
      schema: { type: integer, format: int32 }
...
    VoteType:
      name: voteType
      in: query
      required: true
      schema: { $ref: "#/components/schemas/VoteType" }
    Pageable:
      name: pageable
      in: query
      required: false
      description: |
        Standard Spring Data pagination. Supply as separate query parameters:
        `page` (0-based), `size` (max 100), and `sort` as `property,asc|desc` (repeatable).
      style: form
      explode: true
      schema:
        type: object
        properties:
          page: { type: integer, minimum: 0, default: 0 }
          size: { type: integer, minimum: 1, maximum: 100, default: 20 }
          sort:
            type: array
            items: { type: string }
  responses:
    User:
      description: OK
      content:
        application/json:
          schema: { $ref: "#/components/schemas/UserResponse" }
    UserPage:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/PageResponse" }
              - type: object
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/UserResponse" }
    Post:
      description: OK
      content:
        application/json:
          schema: { $ref: "#/components/schemas/PostResponse" }
    PostPage:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/PageResponse" }
              - type: object
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/PostResponse" }
    Comment:
      description: OK
      content:
        application/json:
          schema: { $ref: "#/components/schemas/CommentResponse" }
    CommentPage:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/PageResponse" }
              - type: object
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/CommentResponse" }
    Subject:
      description: OK
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SubjectResponse" }
...
    File:
      description: OK
      content:
        application/json:
          schema: { $ref: "#/components/schemas/FileResponse" }
    FilePage:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/PageResponse" }
              - type: object
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/FileResponse" }
...
    NotFound:
      description: The addressed resource does not exist
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    Conflict:
      description: Uniqueness, referential or optimistic-locking conflict
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    ValidationFailed:
      description: One or more fields failed validation
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ValidationProblemDetail" }
    StorageFailure:
      description: The object store rejected or failed the operation
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
  schemas:
    VoteType:
      type: string
      enum: [UPVOTE, DOWNVOTE]
    PageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, first, last]
      properties:
        content: { type: array, items: {} }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }
    CreateUserRequest:
      type: object
      required: [username, password, firstName, lastName, email, departmentId, yearId, semesterId]
      properties:
        username: { type: string, maxLength: 40 }
        password: { type: string, minLength: 8, maxLength: 200 }
        firstName: { type: string, maxLength: 60 }
        lastName: { type: string, maxLength: 60 }
        email: { type: string, format: email, maxLength: 180 }
        departmentId: { type: integer, format: int32 }
        yearId: { type: integer, format: int32 }
        semesterId: { type: integer, format: int32 }
        acceptedTermsOfService: { type: boolean, default: false }
    UpdateUserRequest:
      type: object
      required: [username, firstName, lastName, email, departmentId, yearId, semesterId]
      properties:
        username: { type: string, maxLength: 40 }
        password:
          type: [string, "null"]
          minLength: 8
          maxLength: 200
          description: Omit or send null to leave the stored password unchanged.
        firstName: { type: string, maxLength: 60 }
        lastName: { type: string, maxLength: 60 }
        email: { type: string, format: email, maxLength: 180 }
        departmentId: { type: integer, format: int32 }
        yearId: { type: integer, format: int32 }
        semesterId: { type: integer, format: int32 }
        acceptedTermsOfService: { type: [boolean, "null"] }
        canPost: { type: [boolean, "null"] }
        canComment: { type: [boolean, "null"] }
    CreatePostRequest:
      type: object
      required: [title, description, content]
      properties:
        title: { type: string, maxLength: 150 }
        description: { type: string, maxLength: 400 }
        content: { type: string }
    CommentBody:
      type: object
      required: [content]
      properties:
        content: { type: string, maxLength: 5000 }
    SubjectBody:
      type: object
      required: [name, code, departmentId, yearId, semesterId]
      properties:
        name: { type: string, maxLength: 150 }
        code: { type: string, maxLength: 20 }
        lab: { type: boolean, default: false }
        theory: { type: boolean, default: false }
        departmentId: { type: integer, format: int32 }
        yearId: { type: integer, format: int32 }
        semesterId: { type: integer, format: int32 }
    DepartmentResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        name: { type: string }
    YearResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        value: { type: string }
    SemesterResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        name: { type: string }
    RoleResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        name: { type: string }
    UserResponse:
      type: object
      description: The password is never included in a response.
      properties:
        id: { type: integer, format: int32 }
        username: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, format: email }
        acceptedTermsOfService: { type: boolean }
        verified: { type: boolean }
        canPost: { type: boolean }
        canComment: { type: boolean }
        role: { $ref: "#/components/schemas/RoleResponse" }
        department: { $ref: "#/components/schemas/DepartmentResponse" }
        year: { $ref: "#/components/schemas/YearResponse" }
        semester: { $ref: "#/components/schemas/SemesterResponse" }
        dateCreated: { type: string, format: date-time }
        dateLastInteracted: { type: string, format: date-time }
    PostResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        title: { type: string }
        description: { type: string }
        content: { type: string }
        upvoteCount: { type: integer, format: int64 }
        downvoteCount: { type: integer, format: int64 }
        favouriteCount: { type: integer, format: int64 }
        authorId: { type: integer, format: int32 }
        authorUsername: { type: string }
        dateCreated: { type: string, format: date-time }
        dateModified: { type: string, format: date-time }
    CommentResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        content: { type: string }
        upvoteCount: { type: integer, format: int64 }
        downvoteCount: { type: integer, format: int64 }
        postId: { type: integer, format: int32 }
        authorId: { type: integer, format: int32 }
        authorUsername: { type: string }
        dateCreated: { type: string, format: date-time }
        dateModified: { type: string, format: date-time }
    SubjectResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        name: { type: string }
        code: { type: string }
        lab: { type: boolean }
        theory: { type: boolean }
        department: { $ref: "#/components/schemas/DepartmentResponse" }
        year: { $ref: "#/components/schemas/YearResponse" }
        semester: { $ref: "#/components/schemas/SemesterResponse" }
    FileResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        name: { type: string, description: Original client-supplied filename. }
        objectName: { type: string, description: Generated object key in the bucket. }
        objectURL: { type: string, format: uri }
        size: { type: integer, format: int64 }
        favouriteCount: { type: integer, format: int64 }
        subjectId: { type: integer, format: int32 }
        dateCreated: { type: string, format: date-time }
        dateModified: { type: string, format: date-time }
...
    ProblemDetail:
      type: object
      description: RFC 9457 problem detail.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        timestamp: { type: string, format: date-time }
    ValidationProblemDetail:
      allOf:
        - { $ref: "#/components/schemas/ProblemDetail" }
        - type: object
          properties:
            errors:
              type: object
              description: Field name to failure message.
              additionalProperties: { type: string }