# src/main/java/com/documan/dto/response/RoleResponse.java
// MIT License
//
// Copyright (C) 2024 Kalyan Mudumby / @theinhumaneme
//
// Permission is granted to use, copy, modify, merge, publish, distribute,
// sublicense, and/or sell copies of the software.
package com.documan.dto.response;

public record RoleResponse(Integer id, String name) {}